requirements.txt
openai>=1.0.0
rich>=13.0.0
duckduckgo-search>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0

